Update runtime to GNOME 3.38
Runtimes later than GNOME 3.32 allow dconf access, so this should fix
the problem of settings not persisting.

The 3.38 runtime no longer contains enchant, a dependency of gtkspell,
so we have to build it. It also no longer contains the modplug element
in GStreamer, so we have to build that (and delete everything else so
that we install only the modplug element.)

Closes: #1
Closes: #4