Add missing file system permissions
It's unfortunate that we have to give permission to access the whole
documents folder, but as far as I can tell, it's not possible for
select-folder dialogs to work with the desktop portal. Even after
porting to GTK 3 it still didn't work unless using
GtkFileChooserNative, which is not available in GTK 2.

(It may still be that we need to give access to the user's home
folder, but that can be changed later if need be.)